refactor: consolidate gallery mutation state tails
