--- src/main/services/openaiImage.ts
@@ -1,1 +1,0 @@
-export * from "./openaiImageAdapter.js";
